refactor(memory): 将 MemoryItem 抽取到独立文件 item.py
- 新建 api/memory/manager/item.py：MemoryItem 数据模型 + _now() 辅助函数
- base.py：删除 MemoryItem 定义，改为从 item.py 导入；更新类文档（原语说明、CRUD 覆写指南、工具方法清单）
- yaml.py：修正 _write_lock 返回值标注 Generator
- manager/__init__.py：导入路径从 yaml.py 改为 item.py
