+++ api/memory/manager/__init__.py
@@ -2,9 +2,9 @@
 
 from api.memory.manager.base import BaseMemoryManager
 from api.memory.manager.builder import MemoryManagerBuilder
+from api.memory.manager.item import MemoryItem
 from api.memory.manager.yaml import (
     MAX_DESC_LENGTH,
-    MemoryItem,
     YamlMemoryManager,
 )
